+++ services/web/server/src/simcore_service_webserver/users/plugin.py
@@ -13,7 +13,7 @@ from ..user_notifications.bootstrap import (
 )
 from ..user_preferences.bootstrap import setup_user_preferences_feature
 from ..user_tokens.bootstrap import setup_user_tokens_feature
-from ._controller.rest import accounts_rest, users_rest
+from ._controller.rest import accounts_rest, profile_rest
 
 _logger = logging.getLogger(__name__)
 
@@ -31,7 +31,7 @@ def setup_users(app: web.Application):
     assert app[APP_SETTINGS_APPKEY].WEBSERVER_USERS  # nosec
     setup_observer_registry(app)
 
-    app.router.add_routes(users_rest.routes)
+    app.router.add_routes(profile_rest.routes)
     app.router.add_routes(accounts_rest.routes)
 
     setup_user_notification_feature(app)
